WIP chore: add `CBBTCSTAGE/base-ethereum-superseed`
