+++ typescript/infra/config/environments/mainnet3/warp/warpIds.ts
@@ -53,7 +53,9 @@ export enum WarpRouteIds {
   ArbitrumBaseBlastBscEthereumGnosisLiskMantleModeOptimismPolygonScrollZeroNetworkZoraMainnet = 'ETH/arbitrum-base-blast-bsc-ethereum-gnosis-lisk-mantle-mode-optimism-polygon-scroll-zeronetwork-zoramainnet',
   AppchainBaseUSDC = 'USDC/appchain-base',
   BobaBsquaredSwellUBTC = 'UBTC/boba-bsquared-swell',
-  EthereumSuperseedCBBTC = 'CBBTC/ethereum-superseed',
+  // TODO: uncomment after merging the staging route to registry
+  // BaseEthereumSuperseedCBBTC = 'CBBTC/base-ethereum-superseed',
+  BaseEthereumSuperseedCBBTCSTAGE = 'CBBTCSTAGE/base-ethereum-superseed',
   EthereumSuperseedUSDC = 'USDC/ethereum-superseed',
   EthereumFormUSDT = 'USDT/ethereum-form',
   EthereumFormUSDC = 'USDC/ethereum-form',
